add required packages for expo-routing
