+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "coop-word-solver-ts",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
@@ -20,6 +20,7 @@
     "expo-dev-client": "~5.2.4",
     "expo-device": "~7.1.4",
     "expo-image": "~2.3.2",
+    "expo-linking": "~7.1.7",
     "expo-notifications": "~0.31.4",
     "expo-router": "~5.1.4",
     "expo-splash-screen": "~0.30.10",
